Resources & Renaming
move get_false_dup_genes_path() to v4 release resources, and add a renaming dict for globals
